Audit library refactor: split lib/audits/ from scripts/audit-*.ts
Mechanical refactor. The 5 audit scripts shipped on this branch all
had their logic inline in scripts/audit-*.ts, which means the
not-yet-written cron handlers would need to shell out to `tsx
scripts/...` (flaky in a serverless bundle that may not include
tsx or the source file under scripts/). Split each into a library
module with a typed run() function + a thin CLI wrapper.

- lib/audits/auth.ts → runAuditAuth() { total, rows, grouped, unauthed }
- lib/audits/dynamic.ts → runAuditDynamic() { total, dbReading, staticOk, rows, violations }
- lib/audits/tenant.ts → runAuditTenant() { schema, routesScanned, violations, ok, schemaErrors }
- lib/audits/jurisdiction.ts → runAuditJurisdiction() { totalFiles, hits, nonAllowlisted, perTerm }
- lib/audits/kb-staleness.ts → runAuditKbStaleness({ now?, root? }) { now, total, counts, articles, badCount }

Injectable clock on kb-staleness lets the weekly cron be
deterministically tested without freezing system time.

Zero behaviour change for the CLI gate: every script prints
identical output and exits identically. Re-verified A-002 / A-003 /
A-004 / A-005 / A-010 / A-011.

Next commit: /api/cron/* handlers that call these runAuditX()
functions inside runCron.
